demo_mcq/requirements.txt
# ─── Core ──────────────────────────────────────────────────────
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0

# ─── NLP / QA Generation ───────────────────────────────────────
# torch: cài theo hướng dẫn https://pytorch.org/get-started/locally/
torch>=2.1.0
transformers==4.51.3
sentencepiece>=0.1.99
accelerate>=0.27.0

# ─── LLM Backends ──────────────────────────────────────────────
openai>=1.12.0
google-generativeai>=0.4.0

# ─── Export Word / PDF ─────────────────────────────────────────
python-docx>=1.1.0
fpdf2>=2.7.9
